chore: relocate remaining capabilities to _core and rename signal.py to signal_binding.py to avoid confusion with related module and better name the content.
