+++ src/bootstack/widgets/_impl/composites/tabs/tabitem.py
@@ -13,7 +13,7 @@ from bootstack.widgets._impl.primitives.button import Button
 from bootstack.widgets._impl.primitives.label import Label
 from bootstack.widgets._impl.mixins import configure_delegate
 from bootstack.widgets.types import Master
-from bootstack._core.capabilities.signals import normalize_signal
+from bootstack._core.signal_binding import normalize_signal
 
 if TYPE_CHECKING:
     from bootstack.signals import Signal
